# apps/core/docs/swagger.yaml
basePath: /api/v1
definitions:
...
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.AddressItem:
    properties:
      address:
        type: string
      city:
        type: string
      district:
        type: string
      id:
        type: integer
      is_default:
        type: boolean
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      province:
        type: string
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.AddressListResponse:
    properties:
      addresses:
        items:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.AddressItem'
        type: array
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.CreateAddressRequest:
    properties:
      address:
        maxLength: 255
        type: string
      city:
        maxLength: 50
        type: string
      district:
        maxLength: 50
        type: string
      is_default:
        type: boolean
      name:
        maxLength: 50
        type: string
      phone:
        maxLength: 20
        type: string
      postal_code:
        maxLength: 20
        type: string
      province:
        maxLength: 50
        type: string
    required:
    - address
    - name
    - phone
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.NotificationSettings:
    properties:
      email_enabled:
        type: boolean
      push_enabled:
        type: boolean
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.PrivacySettings:
    properties:
      is_public:
        type: boolean
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.ProfileResponse:
    properties:
      avatar_url:
        type: string
      intro:
        type: string
      location:
        type: string
      nickname:
        type: string
      user_id:
        type: integer
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.UpdateAddressRequest:
    properties:
      address:
        type: string
      city:
        type: string
      district:
        type: string
      is_default:
        type: boolean
      name:
        type: string
      phone:
        type: string
      postal_code:
        type: string
      province:
        type: string
    type: object
  github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.UpdateProfileRequest:
    properties:
      avatar_url:
        type: string
      intro:
        type: string
      location:
        type: string
      nickname:
        type: string
    type: object
...
host: localhost:8080
info:
  contact:
    email: support@swagger.io
    name: API Support
    url: http://www.swagger.io/support
  description: This is the core backend service for RevieU
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  termsOfService: http://swagger.io/terms/
  title: RevieU Core API
  version: "1.0"
paths:
...
  /user/account:
    delete:
      consumes:
      - application/json
      description: Schedules account deletion (cooling period). Due deletions are
        executed asynchronously by a background worker.
      parameters:
      - description: Deletion reason
        in: body
        name: request
        schema:
          additionalProperties:
            type: string
          type: object
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request account deletion
      tags:
      - user
  /user/account/export:
    post:
      description: Queues a user data export
      produces:
      - application/json
      responses:
        "202":
          description: Accepted
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Request account export
      tags:
      - user
  /user/addresses:
    get:
      description: Returns the authenticated user's saved addresses
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.AddressListResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List addresses
      tags:
      - user
    post:
      consumes:
      - application/json
      description: Adds a new address for the authenticated user
      parameters:
      - description: Create Address Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.CreateAddressRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.AddressItem'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Create address
      tags:
      - user
  /user/addresses/{id}:
    delete:
      description: Deletes an address
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Delete address
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Updates an existing address
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      - description: Update Address Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.UpdateAddressRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update address
      tags:
      - user
  /user/addresses/{id}/default:
    post:
      description: Sets an address as default
      parameters:
      - description: Address ID
        in: path
        name: id
        required: true
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Set default address
      tags:
      - user
...
  /user/followers:
    get:
      description: Returns followers of the authenticated user
      parameters:
      - description: Cursor
        in: query
        name: cursor
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List followers
      tags:
      - user
  /user/following/merchants:
    get:
      description: Returns merchants the authenticated user follows
      parameters:
      - description: Cursor
        in: query
        name: cursor
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List following merchants
      tags:
      - user
  /user/following/users:
    get:
      description: Returns users the authenticated user follows
      parameters:
      - description: Cursor
        in: query
        name: cursor
        type: integer
      - description: Limit
        in: query
        name: limit
        type: integer
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties: true
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: List following users
      tags:
      - user
...
  /user/notifications:
    get:
      description: Returns the authenticated user's notification settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.NotificationSettings'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get notification settings
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's notification settings
      parameters:
      - description: Notification Settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.NotificationSettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update notification settings
      tags:
      - user
...
  /user/privacy:
    get:
      description: Returns the authenticated user's privacy settings
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.PrivacySettings'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get privacy settings
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Updates the authenticated user's privacy settings
      parameters:
      - description: Privacy Settings
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.PrivacySettings'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update privacy settings
      tags:
      - user
  /user/profile:
    get:
      description: Returns the authenticated user's profile
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.ProfileResponse'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Get current user profile
      tags:
      - user
    patch:
      consumes:
      - application/json
      description: Updates nickname, avatar, intro, or location
      parameters:
      - description: Update Profile Request
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/github_com_revieu-corp_revieu-core-api-go_apps_core_internal_domain_user_dto.UpdateProfileRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            additionalProperties:
              type: string
            type: object
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      summary: Update current user profile
      tags:
      - user
...
schemes:
- http
- https
swagger: "2.0"